Update librtlsdr.c
I was working with librtlsdr, and making continuous changes to sample_rate.  The library kept echoing back to the console when the set rate was only decimals away from the actual rate.  I was piping to another program that was writing its output to the console, and this constant feedback was scrolling it off the screen.

So I changed the librtlstd.c program to have a global variable called Debug_Librtlsdr.  I then wrapped all the fprintf messages.  When this variable is set to 0, nothing prints.  When 1, messages I consider errors print.  When 2, info messages as well as errors print.  I consider the warning about the frequency to be an info message.

I wrote a little function that can be called to set the Debug_Librtlsdr variable.  

And so, here we are.