Make librtlsdr cmake subdirectory friendly
Change CMAKE_SOURCE_DIR to PROJECT_SOURCE_DIR to allow for inclusion of
librtlsdr as a subdirectory in another cmake project. For a standalone build,
PROJECT_SOURCE_DIR and CMAKE_SOURCE_DIR work pretty much the same, but
if a projet is included into another with add_subdirectory PROJECT_SOURCE_DIR
must be used.